backend/requirements.txt
# Core Framework
Flask==3.0.0
Flask-CORS==4.0.0
livekit==1.0.17
livekit-agents==1.2.17
livekit-plugins-openai==0.7.0
livekit-plugins-silero==0.6.0
livekit-api==1.0.7

openai==1.101.0

# Database
SQLAlchemy==2.0.23

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Date & Time
python-dateutil==2.8.2

# Networking
requests==2.31.0

# Config
python-dotenv==1.0.0

# Validation
jsonschema==4.20.0

# Logging
colorlog==6.8.0
mysql-connector-python
